pyproject.toml
[tool.poetry]
name = "cdek-bot"
version = "0.1.0"
description = "This bot using for detecting bar codes in CDEK's form and print it on stickers"
authors = ["Loginov <itsmeegorloginov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
telebot = "^0.0.5"
pillow = "^10.3.0"
python-barcode = "^0.15.1"
pyzbar = "^0.1.9"
opencv-python = "^4.9.0.80"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
